Implement queueing and duplicate manager classes, and a packet egressing thread
A packet queuing and a packet duplicate managing class have been implemented
with thread safety using mutexes and conditional variables. This moves much of
the logic out of the packet capturing function and enables the project for
multithreading.

The packet egressing thread has also been implemented, which waits on the packet
queue for frames to broadcast and further splits up the logic previously
consolidated in the packet capturing function/pcpp threads.