Add a processing step to the PacketQueue pipeline
This new step, implemented via the function process_packet(), goes in
between the push and pop functions of the queue. A new thread has been
added which calls this function repeatedly to add metadata to queue
entries. It does not do this yet, but it must be called before the packet
corresponding to the entry can be egressed.

The function which initially receives packets and pushes them to the queue
has been renamed to receive_packet() (from process_packet()) to avoid
confusion with the new processing threads.