Create the MacAddressTable class to maintain MAC address to intf mappings
This class contains thread safe functions to add mappings, query the
current mapping table, and to age old mappings out. It uses std::map behind
the scenes, mapping a given MAC address to an interface and an internal
timestamp. The timestamp is used to determine if an entry should be removed
based on a fixed maximum length.