Integrate MAC address learning into the packet queueing pipeline
MAC address to interface mappings are now maintained throughout the
switch's runtime using MacAddrTable, and the contents of the table are
used in the processing step of packet pipeline to transmit frames to a
single interface if its destination MAC address matches an entry in the
table.

Unrecognized destination MAC addresses still result in broadcasted frames,
but the logic for doing so has been moved from the egressing thread to the
packet processing function in PacketQueue.