Implement CLI parser and interpreter
This implements a basic parser using Flex, which recognizes a handful of
tokens and outputs a value corresponding to it. The function cli() uses the
generated parser to build a vector of tokens which make up the user's
commmand and passes it to the interpret() function of the new class
CliInterpreter.

CliInterpreter maintains a tree of all valid CLI commands and the functions
which should be called when a CLI command is made. The call to interpret()
walks down the tree and calls the function corresponding to input, assuming
it is valid. If it was invalid, an error code is returned.

CMakeLists.txt and .gitignore have been updated to accommodate the
addition of Flex.