Create a script which automates testing infrastructure creation
This script is very similar to init_hosts_switch.sh in how it is
structured, where instead of creating many hosts and a single vswitch, it
creates a single testing container and a single vswitch.

Both containers each have the same number of interfaces, with virtual links
between each pair. In the future, the testing container will run a program
which transmits traffic to the vswitch and verifies that it received the
correct packets in response.