Implement entry point for the program to run on the testing container
This program takes a single command line argument specifying which test
should be run. That argument is used to decide which test setup function
should be called, which determines which packets should be sent and
expected to be received.

Once this is complete, the test is run and its result it returned to the
caller.