Implement a test orchestrator
This new program automatically, for each test defined for the vswitch,
starts the vswitch program on the vswitch container with a given CLI
configuration and starts the vswitch_testing program on the vswitch-testing
container. It records the results of each test and outputs them once they
have all completed.

Note that a new instance of the vswitch program is started and ended for
each test run. This is done to prevent cross-contamination of test results.