Add the "show interfaces" CLI command
This uses the fork() and execvp() calls with the program `ip` to show
information regarding the interfaces managed by the switch.

The opening and closing messages when entering and exiting the program
have been removed as this command replaces their purpose, and the lexer
has been changed to recognize the string "interfaces" rather than
"interface" for the INTF token, which better matches the CLIs of commercial
switches.